qa_agent/requirements.txt
playwright>=1.44.0
anthropic>=0.29.0
pyyaml>=6.0.1
